+++ python/pathway/io/mysql/__init__.py
@@ -10,9 +10,7 @@ from pathway.internals.expression import ColumnReference
 from pathway.internals.runtime_type_check import check_arg_types
 from pathway.internals.table import Table
 from pathway.internals.trace import trace_user_frame
-from pathway.io._utils import init_mode_from_str
-
-_SNAPSHOT_OUTPUT_TABLE_TYPE = "snapshot"
+from pathway.io._utils import SNAPSHOT_OUTPUT_TABLE_TYPE, init_mode_from_str
 
 
 @check_arg_types
@@ -160,7 +158,7 @@ for MySQL database.
         max_batch_size=max_batch_size,
         table_name=table_name,
         table_writer_init_mode=init_mode_from_str(init_mode),
-        snapshot_maintenance_on_output=output_table_type == _SNAPSHOT_OUTPUT_TABLE_TYPE,
+        snapshot_maintenance_on_output=output_table_type == SNAPSHOT_OUTPUT_TABLE_TYPE,
     )
 
     key_field_names = None
@@ -175,7 +173,7 @@ for MySQL database.
     )
 
     datasink_type = (
-        "snapshot" if output_table_type == _SNAPSHOT_OUTPUT_TABLE_TYPE else "sink"
+        "snapshot" if output_table_type == SNAPSHOT_OUTPUT_TABLE_TYPE else "sink"
     )
     table.to(
         datasink.GenericDataSink(